feat(Contracts): setActive & seperate table,tableData
